AI_Services/app/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
sqlalchemy==2.0.42
psycopg2-binary==2.9.10
python-dotenv==1.1.1
pydantic-settings==2.10.1
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
email-validator==2.2.0
langgraph==0.2.50
langchain==0.3.27
langchain-google-genai==2.0.10
google-generativeai==0.8.5
pinecone==5.4.2
httpx==0.28.1
jinja2==3.1.6
python-multipart==0.0.20
PyPDF2==2.10.9
